Move conversion functions to their own module
